[doc]  draft docs for branches and lush
Lush.md is the central point to describe the project.
I hope I have now pinned down the main concepts and developped  a semi stable vocabulary.
the start of the lush doc lists all the important concepts and their names. They may need to be explicited later in the doc. But for me it is good to see them at a glance.